Move helpers into ControllerHelpers::Common
Based on point here: https://github.com/spree/spree/pull/2079#issuecomment-9371007
